Rename file resources-nav.vue to ResourcesNav.vue
