+++ components/AppFooter.vue
@@ -84,7 +84,7 @@
 </template>
 
 <script>
-import ResourcesNav from '~/components/resources-nav'
+import ResourcesNav from '@/components/ResourcesNav'
 import AboutNav from '@/components/AboutNav'
 
 export default {
