+++ components/ResourcesNav.vue
